remove audio utils, add yaml reader
